Update de.lang
lol this was hard work :-)
i dont know thats: lower hotpluggin mean, so i dont know if this translation is correct. 
i also dont know whats text_calibrating and where this string is used?
and i did not understand archidroid.tweaks.nocell.desc= ?
deselect if not wanted? (so it is preselected to have bloatware?)